Se cambia tipo de slaida del método suma a long

+++ src/main/java/com/unisys/ejercicios/Ejercicio1.java
@@ -16,7 +16,7 @@ public class Ejercicio1 {
         presupuesto=presu;
         longitud=longit;
     }
-    public double suma ()
+    public long suma ()
     {
         return gasol+marianico;
     }
